fix: proper suspense in nav header
